+++ src/imuposer/config.py
@@ -40,9 +40,9 @@ class Config:
         self.raw_amass_path = self.root_dir / "data/raw/AMASS"
 
         self.processed_dip_path = self.root_dir / "data/processed/DIP_IMU"
-        self.processed_like_transpose_path = self.root_dir / "data/processed_transpose"
-        self.processed_imu_poser = self.root_dir / "data/processed_imu_poser"
-        self.processed_imu_poser_25fps = self.root_dir / "data/processed_imu_poser_25fps"
+        self.processed_like_transpose_path = self.root_dir / "data/processed"
+        self.processed_imu_poser = self.root_dir / "data/processed_imuposer"
+        self.processed_imu_poser_25fps = self.root_dir / "data/processed_imuposer_25fps"
         self.processed_uwb_poser_5fps = self.root_dir / "data/processed_uwb_poser_5fps"
 
         self.vposer_ckpt_path = self.root_dir / "extern/vposer_v2_05"
